fix(query-bar): move helpers to utils file
