+++ src/common/utils.js
@@ -1,0 +1,9 @@
+// @flow
+import isFinite from 'lodash/isFinite';
+
+const isFloat = (value: number) => {
+    return !/^\s*$/.test(value) && isFinite(value);
+};
+
+// eslint-disable-next-line import/prefer-default-export
+export { isFloat };
