--- src/features/query-bar/validator.js
@@ -1,28 +1,0 @@
-// @flow
-import { isInteger, isNaN } from 'lodash';
-
-import { FLOAT } from './constants';
-
-const isFloat = (value: string) => {
-    return !/^\s*$/.test(value) && !isNaN(Number(value));
-};
-
-const isInt = (value: string) => {
-    return isInteger(Number(value)) && !isNaN(Number(value));
-};
-
-const isValidValue = (type: string, value?: string | Date | any) => {
-    if (type === FLOAT) {
-        if (value != null && isFloat(String(value))) {
-            return true;
-        }
-        if (value == null) {
-            return true;
-        }
-        return false;
-    }
-    return false;
-};
-
-// eslint-disable-next-line import/prefer-default-export
-export { isFloat, isInt, isValidValue };
